Refactor to allow ignoring Config trait
