+++ .github/workflows/main.yml
@@ -192,7 +192,7 @@ jobs:
         run: |
           cargo test-all-features
       - name: 🙊 Code Coverage (vergen) 🙊
-        run: cargo tarpaulin --config tarpaulin.toml --exclude-files src/lib.rs --out Xml --skip-clean --target-dir target/coverage
+        run: cargo +nightly tarpaulin --config tarpaulin.toml --exclude-files src/lib.rs src/config.rs target/* --out Xml --skip-clean --target-dir target/coverage
       - name: 🙊 Upload Code Coverage 🙊
         uses: codecov/codecov-action@v3
         with:
