+++ src/config.rs
@@ -10,7 +10,6 @@ use tracing_subscriber::fmt::format::FmtSpan;
 
 /// Implement this trait to supply tracing configuration that can be used to build a [`Layer`](tracing_subscriber::Layer)
 /// with functions such as [`full_filtered`](crate::full_filtered).
-#[cfg_attr(tarpaulin, ignore)]
 pub trait Config {
     /// Get the quiet count (these are normally pulled from the command line arguments)
     fn quiet(&self) -> u8;
@@ -70,94 +69,5 @@ pub trait Config {
     /// This defaults to false
     fn with_thread_names(&self) -> bool {
         false
-    }
-}
-
-#[doc(hidden)]
-#[derive(Clone, Copy, Debug)]
-pub struct TestAll;
-
-impl Config for TestAll {
-    fn quiet(&self) -> u8 {
-        0
-    }
-
-    fn verbose(&self) -> u8 {
-        3
-    }
-
-    #[cfg(feature = "json")]
-    fn with_current_span(&self) -> bool {
-        true
-    }
-
-    fn with_file(&self) -> bool {
-        true
-    }
-
-    fn with_line_number(&self) -> bool {
-        true
-    }
-
-    fn with_span_events(&self) -> Option<FmtSpan> {
-        Some(FmtSpan::FULL)
-    }
-
-    #[cfg(feature = "json")]
-    fn with_span_list(&self) -> bool {
-        true
-    }
-
-    fn with_target(&self) -> bool {
-        true
-    }
-
-    fn with_thread_ids(&self) -> bool {
-        true
-    }
-
-    fn with_thread_names(&self) -> bool {
-        true
-    }
-}
-
-#[cfg(test)]
-pub(crate) mod test {
-    use super::Config;
-
-    #[derive(Clone, Debug)]
-    pub(crate) struct TestConfig;
-
-    impl Config for TestConfig {
-        fn quiet(&self) -> u8 {
-            0
-        }
-
-        fn verbose(&self) -> u8 {
-            1
-        }
-    }
-
-    #[derive(Clone, Debug)]
-    pub(crate) struct TestJson;
-
-    impl Config for TestJson {
-        fn quiet(&self) -> u8 {
-            0
-        }
-
-        fn verbose(&self) -> u8 {
-            1
-        }
-
-        #[cfg(feature = "json")]
-        fn with_current_span(&self) -> bool {
-            true
-        }
-
-        #[cfg(feature = "json")]
-        fn with_span_list(&self) -> bool {
-            true
-        }
     }
 }
