+++ src/format/compact.rs
@@ -97,7 +97,7 @@ mod test {
 
     use super::filtered as compact_filtered;
 
-    use crate::{config::test::TestConfig, set_default, TestAll};
+    use crate::{utils::test::TestConfig, set_default, TestAll};
 
     #[test]
     fn compact_filtered_works() {
