+++ src/format/pretty.rs
@@ -95,7 +95,7 @@ mod test {
 
     use super::filtered as pretty_filtered;
 
-    use crate::{config::test::TestConfig, set_default, TestAll};
+    use crate::{utils::test::TestConfig, set_default, TestAll};
 
     #[test]
     fn pretty_filtered_works() {
